"nuovi_positivi" reverted back instead of "variazione_totale_positivi"
After several analysis, I came to conclusion that "nuovi_positivi" is actually the closest, more precise way to track down individual new cases for ratio computing. The problem is that hospitalyzed and confined patients are cumulative values, and deltas are not reported. Instead, we can obtain a decent precision by adding deaths and recovers to the cumulative value of current positive patients.